Python-Backend-API/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv 
pydantic-settings
google-genai
